refactor: 이미지 검증 함수 분리 및 상부 분리
