+++ src/constants/challenge-post.constants.ts
@@ -35,3 +35,8 @@ export const DATABASE = {
     CREATOR_ID: 'creator_id'
   }
 };
+
+export const FILES = {
+  ALLOWED_TYPES: ['image/png', 'image/jpeg'],
+  MAX_SIZE: 3 * 1024 * 1024
+};
